projects/sre/backups_snapshots/pyproject.toml
[tool.poetry]
name = "backups-snapshots"
version = "0.1.0"
description = ""
authors = ["ekollarova <ekollarova@blizzard.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-compute = "^1.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
